Move BaseHook, BaseOperator, and BaseSensorOperator imports to version_compat for Alibaba provider Airflow 3.0 compatibility
