+++ providers/alibaba/src/airflow/providers/alibaba/cloud/hooks/oss.py
@@ -27,11 +27,7 @@ import oss2
 from oss2.exceptions import ClientError
 
 from airflow.exceptions import AirflowException
-
-try:
-    from airflow.sdk import BaseHook
-except ImportError:
-    from airflow.hooks.base import BaseHook  # type: ignore[attr-defined,no-redef]
+from airflow.providers.alibaba.version_compat import BaseHook
 
 if TYPE_CHECKING:
     try:
