+++ providers/alibaba/src/airflow/providers/alibaba/cloud/operators/analyticdb_spark.py
@@ -24,12 +24,7 @@ from typing import TYPE_CHECKING, Any
 
 from airflow.exceptions import AirflowException
 from airflow.providers.alibaba.cloud.hooks.analyticdb_spark import AnalyticDBSparkHook, AppState
-from airflow.providers.alibaba.version_compat import AIRFLOW_V_3_0_PLUS
-
-if AIRFLOW_V_3_0_PLUS:
-    from airflow.sdk import BaseOperator
-else:
-    from airflow.models import BaseOperator
+from airflow.providers.alibaba.version_compat import BaseOperator
 
 if TYPE_CHECKING:
     from airflow.utils.context import Context
