+++ providers/alibaba/src/airflow/providers/alibaba/cloud/operators/maxcompute.py
@@ -24,12 +24,7 @@ from typing import TYPE_CHECKING
 
 from airflow.providers.alibaba.cloud.hooks.maxcompute import MaxComputeHook
 from airflow.providers.alibaba.cloud.links.maxcompute import MaxComputeLogViewLink
-from airflow.providers.alibaba.version_compat import AIRFLOW_V_3_0_PLUS
-
-if AIRFLOW_V_3_0_PLUS:
-    from airflow.sdk import BaseOperator
-else:
-    from airflow.models import BaseOperator
+from airflow.providers.alibaba.version_compat import BaseOperator
 
 if TYPE_CHECKING:
     from odps.models import Instance
